Add data to json files
