+++ app.rb
@@ -4,6 +4,7 @@ require './book'
 require './rental'
 require './student'
 require './teacher'
+require 'json'
 
 class App
   def initialize
@@ -110,21 +111,4 @@ class App
     end
   end
 
-  def write_json
-    File.open('books.json', 'w') do |file|
-      file.write(@books.to_json)
-    end
-    File.open('people.json', 'w') do |file|
-      file.write(@people.to_json)
-    end
-    File.open('rentals.json', 'w') do |file|
-      file.write(@rentals.to_json)
-    end
-  end
-
-  def read_json
-    @books = JSON.parse(File.read('books.json'))
-    @people = JSON.parse(File.read('people.json'))
-    @rentals = JSON.parse(File.read('rentals.json'))
-  end
 end
