+++ data/books.json
@@ -1,0 +1,7 @@
+[
+  {
+    "json_class": "Book",
+    "title": "Purpose Driven Development",
+    "author": "Eric Evans",
+  }
+]
