+++ data/people.json
@@ -1,0 +1,16 @@
+[
+  {
+    "json_class": "Student",
+    "id": 1,
+    "name": "John",
+    "age": "20",
+    "parent_permission": true
+  },
+  {
+    "json_class": "Teacher",
+    "id": 2,
+    "name": "Jane",
+    "age": "30",
+    "specialization": "Math Teacher"
+  }
+]
